Resolve the last 7 signature disagreements by hand
These were refused by the fixer on purpose: all seven are type WIDTH
changes rather than signedness, and narrowing or widening a field is a
judgement about what the code means. Decided with the maintainer:

  temp_set, climate_set, rocky_set, region_set, alt_put_detail
      uint8_t -> int. The mangled names say H, the callers already pass
      int, and a uint8_t parameter forces a movzx on entry the original
      does not have.
  go_to        uint8_t -> char. D decodes to char, which MSVC signs.
  checksum     uint8_t * -> char *. The body accumulates through the
      pointer, so this changes arithmetic for bytes above 127 - which is
      plausibly why it has not matched.

Two of them split their signature while being fixed, because the
parameter NAMES differ between the header and the definition
(`uint8_t *input` against `char *buffer`; `uint8_t climate` against
`uint8_t rainfall`), so a name-based edit reached only the definition.
C++ compiles a split signature as an overload rather than failing, and
audit_recovered_signatures reads definitions only, so it reported zero
disagreements while two headers still contradicted their own bodies.
Both repaired; all seven now agree in both places.

Corpus-wide disagreements: 7 -> 0 over 2,146 signatures. All 2,523
catalogued locations still extract.

+++ src/map.cpp
@@ -489,7 +489,7 @@ Original Offset: 00591AD0
 Return Value: n/a
 Status: Complete
 */
-void __cdecl temp_set(int x, int y, uint8_t temperature) {
+void __cdecl temp_set(int x, int y, int temperature) {
     Map *tile = map_loc(x, y);
     tile->climate &= 0xF8;
     tile->climate |= temperature & 7;
@@ -511,7 +511,7 @@ Original Offset: 00591A80
 Return Value: n/a
 Status: Complete
 */
-void __cdecl climate_set(int x, int y, uint8_t rainfall) {
+void __cdecl climate_set(int x, int y, int rainfall) {
     Map *tile = map_loc(x, y);
     tile->climate &= 0xE7;
     tile->climate |= (rainfall & 3) << 3;
@@ -599,7 +599,7 @@ Original Offset: 00591260
 Return Value: n/a
 Status: Complete
 */
-void __cdecl alt_put_detail(int x, int y, uint8_t detail) {
+void __cdecl alt_put_detail(int x, int y, int detail) {
     map_loc(x, y)->contour = detail;
 }
 
@@ -653,7 +653,7 @@ Original Offset: 00591B90
 Return Value: n/a
 Status: Complete
 */
-void __cdecl region_set(int x, int y, uint8_t region) {
+void __cdecl region_set(int x, int y, int region) {
     map_loc(x, y)->region = region;
 }
 
@@ -746,7 +746,7 @@ Original Offset: 00591BC0
 Return Value: n/a
 Status: Complete
 */
-void __cdecl rocky_set(int x, int y, uint8_t rocky) {
+void __cdecl rocky_set(int x, int y, int rocky) {
     Map *tile = map_loc(x, y);
     tile->val3 &= 0x3F;
     tile->val3 |= rocky << 6;
